move sims generation to Batch & change sims storage params
