+++ epimodel/gleam/__init__.py
@@ -1,3 +1,2 @@
 from .batch import Batch
 from .definition import GleamDefinition
-from .scenario import generate_simulations
